grsec: Apply grsecurity-2.1.14-2.6.29.6-200907301033.patch
Signature-tree: ccd94d2fd2ec7d9f7db5849cb4816ae5a2eea4d3

+++ arch/sparc/include/asm/uaccess_32.h
@@ -246,7 +246,7 @@ extern unsigned long __copy_user(void __user *to, const void __user *from, unsig
 
 static inline unsigned long copy_to_user(void __user *to, const void *from, unsigned long n)
 {
-	if (((int)n > 0) && __access_ok((unsigned long) to, n))
+	if (((long)n > 0) && __access_ok((unsigned long) to, n))
 		return __copy_user(to, (__force void __user *) from, n);
 	else
 		return n;
@@ -259,7 +259,7 @@ static inline unsigned long __copy_to_user(void __user *to, const void *from, un
 
 static inline unsigned long copy_from_user(void *to, const void __user *from, unsigned long n)
 {
-	if (((int)n > 0) && __access_ok((unsigned long) from, n))
+	if (((long)n > 0) && __access_ok((unsigned long) from, n))
 		return __copy_user((__force void __user *) to, from, n);
 	else
 		return n;
